refactor(client-telemetry): address PR #620 review findings
Split inline_hooks.rs (1,062 lines, 12 hook pairs — past the 700-line
hard cap and the 4-sibling directory threshold) into an inline_hooks/
directory module: mercury_dispatch, engine_frame, engine_loading,
state_flags, anim_notify, console_command, with install orchestration
and the aggregated anchor-pin test in mod.rs. The split also repairs
the misplaced insertion that had orphaned state_field_update_detour's
doc comment between the new detour's attributes.

Align the drop-oracle callee's documented ABI with the code: the
v5-campaign RE pass identifies 0x01590f30 as
EntityDescription_GetExposedClientMethodByIndex returning a
MethodDescription* in EAX, so the detour/trampoline now type the
return as an opaque *mut c_void (was i32) and the doc states the
return instead of implying void.

Reword the "lock-free"/"no locks" hot-path comments: try_emit is a
bounded-channel try_send and the event builder heap-allocates; the
actual constraint is non-blocking drop-on-full with no game-state
access.

Document client.dispatch.method_dropped in the hookpoints catalog
(new "Dispatch drop oracle" section) and correct the stale 11-hook
count in hooks.rs.
